cmd: プロトコルごとにファイルを分割
cmd/execute.go に集約されていた HTTP/MQTT/UDP の実装をプロトコル別の
ファイルに分割し、execute.go はサービス起動のオーケストレーションのみ
担当するようにしました。各ファイルが必要とするインポートだけを保持する
ようになり、責務が明確になります。

- cmd/http.go : startWebServer
- cmd/mqtt.go : startMQTTBroker, mqttAuthHook, mqttTopicMatch ほか
- cmd/udp.go  : startUDPListener, UdpMessage
- cmd/execute.go : execute() のみ
